Fix qemu-img info vhd format issue
Due to the fact that qemu-img seeks the VHD signature at the wrong
offset, fixed size VHD images are probed as 'raw'.

This patch is a workaorund for the issue, adding an extra check
seeking the VHD 'conectix' at the right offset.

This workaround can be removed after the following bug fix merges:
https://patchwork.ozlabs.org/patch/375762/