Fix retrieving the volume path and format when using SMBFS drivers
The SMBFS drivers use the image format in order to construct the
according image path, as some image formats require having the
proper extension.

The format is either retrieved using qemu-img if the image exists,
either via extra specs or config options if the image does not
exist.

The issue is that when searching for local images, the extension is
not used, falling back to extra specs, which can cause lazy load
errors or config options, which might change.

This issue occures when vhd/x images are used. The fix consists in
simply using the possible extensions when trying to find existing
images.