SMBFS: Fix retrieving total allocated size
Currently, the Windows SMBFS driver parses all the VHD/X images
present on the configured share in order to retrieve the allocated
size. Not only that this can be slow when having many volumes, but
it can also fail in certain environements, when the volumes are
attached.

In order to workaround this, a data file is used, placed at a
configurable location. This is a JSON containing allocation sizes
for each volume created by the driver, as well as the total
allocated size.