Add differencing image support for cached images
Using differencing VHD/X images pointing to the cached images
can improve the time and disk space needed in order to create
volumes from images.

This feature can be enabled by using a new config option.

In case of SMB backends, the base images will be cached on the
same share as the dependent images as those will be directly
accessed by the hypervisors.