RemoteFS: Fix the snapshot delete operation
The RemoteFS offline snapshot logic is flawed when it comes to
deleting intermediary snapshots. This leads to unexpected behavior.

Instead of merging the higher image, the actual snapshot image
should be merged into its parent, reconnecting the higher image,
if existing.

Let's take the following image chain for example:

volume1234 -> volume1234.aaaa -> volume.1234.bbbb
   a               b                    c

By creating a new volume using the snapshot bbbb, we should have a
new volume having the content "b", as that was the state of the
volume when the snapshot was taken.

At the moment, when the snapshot aaaa is deleted, volume.1234.bbbb
will be merged into volume1234.aaaa, this becoming the new active
image. This means will get the following chain:

volume1234 -> volume1234.aaaa
              (now representing snap bbbb)
   a              c

Creating a volume using snapshot bbbb will now contain data "a"
instead, thus the same snapshot leads to a different point in time.

This affects all the RemoteFS based drivers which support snapshots.