Fix getting attached VHD images info
This patch fixes the issue by using the "VIRTUAL_DISK_DETACH" access
mask. Note that this does not affect the process using the image.

Citing from the MSFT docs:
"To prevent an open request failure when attempting to open a handle
to a permanently attached virtual disk, the following requirements apply:
The VirtualDiskAccessMask parameter must include the
VIRTUAL_DISK_ACCESS_DETACH (0x00040000) flag.