SMBFS: Fix initialize connection issues caused by in-use images
In some cases, if the volume is in use, attempting to retrieve
image informations will raise permission errors.

This is mainly affecting live migration when having volumes attached
over SMB, when using Hyper-V.

The only information that we attempt to retrieve in this case is
the volume format, which is needed only by Libvirt. This patch will
let the initialize_connection pass even if the format could not be
retrieved. Still, this should not cause issues to Libvirt.