SMBFS: Fix initialize connection issues caused by in-use images
In some cases, if the volume is in use, attempting to retrieve
image informations will raise permission errors.

This is mainly affecting live migration when having volumes attached
over SMB, when using Hyper-V.

The only information that we attempt to retrieve in this case is
the volume format. We can easily avoid these kind of issues by always
using the volume format as an extension for the image files (at
the moment we do this only for vhd/x images).

Change-Id: Id1c713b2d3047c5ee5746207d73059b19ce4563b
Closes-Bug: #1554027

Conflicts:
   cinder/tests/test_smbfs.py