Windows SMBFS: fix VHD/x resize
The os-win vhdutils resize method, by default, treats the requested
size as the maximum VHD/x file size. Considering the size of the
internal metadata, the virtual size of the image will be slightly
smaller.

Resizing an image may fail when the image size is close to the
requested size, in which case the new calculated internal size may be
smaller than the current image internal size.

This fix consists in going back to the old behavior (pre os-win),
ensuring that the new VHD/x internal size always matches the requested
VHD/x size, regardless of the internal metadata overhead (which is
really small anyway).

Change-Id: Ib910ae388e210f33e4cfd87438e8e0ee17020d40
Closes-Bug: #1591988