RemoteFS: enable image volume cache
At the moment, the RemoteFS based drivers cannot leverage the generic
image volume cache feature.

The reason is that this involves cloning a newly created volume while
still being in 'downloading' state, which is an unacceptable state
from the driver's point of view.

This change ensures that we set the original volume status to 'cloning'
before attempting to create the 'image-cache' volume, adding this state
as an acceptable state at the driver level as well.

While this may be seen as an workaround, it does actually help providing
a better overview of the volume status during the creation process.

This change attempts to be as less intrusive as possible at the same
time, not changing the volume status during 'regular' volume clones,
which may be undesirable (causing some other operations to be rejected,
for example volume attachments).

Closes-Bug: #1685277