RemoteFS: enable image volume cache
At the moment, the RemoteFS based drivers cannot leverage the generic
image volume cache feature.

The reason is that this involves cloning a newly created volume while
still being in 'downloading' state, which is an unacceptable state
from the driver's point of view.

This change adds 'downloading' as an acceptable state for the volume
clone operation as well as for creating/deleting snapshots used
internally by the volume clone operation.

Note that 'regular' volume clones will be rejected before reaching
the driver when having a source volume in 'downloading' state.

Change-Id: I84bb3f062637031f7f46b414d6cbbff0bb0292ea
Closes-Bug: #1685277