SMBFS: drop JSON file storing allocation data
Due to the fact that we cannot query in-use images, some time ago
we've started storing the total disk image allocated size within
a JSON file.

At the same time, we're caching those values within the service.

As we're aiming to support A-A deployments, also considering the
risk of having this file corrupted, we're now dropping it.

Instead of using a separate DB, we're just going to rely on the
Cinder DB when retrieving the allocated space for a specific share.

One reason why we did not go with this approach from the beginning
is the fact that we needed support for reporting each share as a pool,
which is accomplished by the other patches from this chain.

Change-Id: I2085c040d7140f8ba685f6a36c4d3b5b30b14103