SMBFS: deprecate provisioning ratio config options
This change deprecates the SMBFS driver oversubmit ratio and used
ratio config options, using the *generic* config options instead.

Note that as per a recent change, the superclass uses the *generic*
options when reporting volume stats. This change ensures that we're
moving to those options in a backwards compatible way.

Change-Id: If2e6b9c4bcdaace68469855f1063efca9c502d09