Double check share capacity when creating volumes
At the moment, we rely on the scheduler to select a share for a
certain volume. There is a small timeframe in which the scheduler
may be out of sync, for which reason we're double checking the
available space when creating volumes.

We'll need to decide if we want this to go upstream and in a driver
independent way.

Change-Id: I8f42022ae786063398cb4ca771a860f532ee2029