smbfs: set VHD UUID using volume UUID
The SMBFS driver stores volumes as VHD/x image files. This change
will set the unique id VHD field using the volume id.

Hyper-V propagates this id to the guest, which allows it to safely
identify the disk. The kubernetes openstack provider will also
benefit from this change.

http://paste.openstack.org/raw/801011/

This operation requires os-win 5.4.0, for which reason we're bumping
the requirement.

Depends-On: Ied73997e6f5f3ded9827703867f059ef3dfca159

Change-Id: I459ff3c8798e218919b6b2f9e1c2200203efeeff